Some modifications and additions to the existing code.
Adds disabling of automtic migration of features.

Modifies reboot code so that all nodes with all roles are rebooted after nodes with master role.

Playbook structure redesigned. For every role  there is a playbook. For example:  Now if task fails at reboot one can simply run the reboot-playbook task.
`playbook-migration.yml` is a list of playbooks for all the migration related roles in sequence.
`playbook-rollback.yml` is a list of playbooks for all the rollback related roles in sequence.

During rollback checks if version is OCP 4.17 and fails the rollback as OpenShiftSDN is removed from this version.